test: :white_check_mark: test ApplePowermetrics
